% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{OSM}
\alias{OSM}
\title{Open Street Map instance initiator}
\usage{
OSM(place_name = NULL)
}
\arguments{
\item{place_name}{defult=NULL, will assume you copy bbox from https://openstreetmap.org}
}
\value{

}
\description{
Open Street Map instance initiator
}
\examples{
none
}
