% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{osm_rename}
\alias{osm_rename}
\title{Resolve the osm_sf in ggplot name error problem}
\usage{
osm_rename(sf_object)
}
\arguments{
\item{sf_object}{An simple feature dataframe exported from osmdata::osm_sf function}
}
\value{

}
\description{
Resolve the osm_sf in ggplot name error problem
}
